--- packages/pymedphys_utilities/src/pymedphys_utilities/types/__init__.py
@@ -1,1 +1,0 @@
-from .core import to_tuple
